Seperated content source into multiple smaller files.
